Implement setting to tootle footer
