+++ src/views/Settings/Settings.vue
@@ -26,7 +26,7 @@
 		<ConfigureCheck />
 		<DownloadBinaries />
 		<RootCertificate />
-		<UrlValidation />
+		<SignPage />
 		<AllowedGroups />
 		<LegalInformation />
 	</NcSettingsSection>
@@ -37,7 +37,7 @@ import NcSettingsSection from '@nextcloud/vue/dist/Components/NcSettingsSection'
 import ConfigureCheck from './ConfigureCheck.vue'
 import DownloadBinaries from './DownloadBinaries.vue'
 import RootCertificate from './RootCertificate.vue'
-import UrlValidation from './UrlValidation.vue'
+import SignPage from './SignPage.vue'
 import AllowedGroups from './AllowedGroups.vue'
 import LegalInformation from './LegalInformation.vue'
 
@@ -48,7 +48,7 @@ export default {
 		ConfigureCheck,
 		DownloadBinaries,
 		RootCertificate,
-		UrlValidation,
+		SignPage,
 		AllowedGroups,
 		LegalInformation,
 	},
